refactor: Board에서 오목 체크 분리
