+++ src/main/kotlin/omok/model/rule/Rule.kt
@@ -1,0 +1,11 @@
+package omok.model.rule
+
+import omok.model.Board
+import omok.model.entity.StoneColor
+
+interface Rule {
+    fun check(
+        board: Board,
+        color: StoneColor,
+    ): Boolean
+}
