+++ src/main/kotlin/omok/model/turn/BlackTurn.kt
@@ -3,12 +3,13 @@ package omok.model.turn
 import omok.model.Board
 import omok.model.entity.Point
 import omok.model.entity.StoneColor
+import omok.model.rule.OmokRule
 
 class BlackTurn(val board: Board) : Turn {
     override fun placeStone(point: Point): Turn {
         board.place(point, StoneColor.BLACK)
 
-        if (board.isFull() || board.startCheckOmok(StoneColor.BLACK)) return Finished(StoneColor.BLACK)
+        if (board.isFull() || OmokRule().check(board, StoneColor.BLACK)) return Finished(StoneColor.BLACK)
 
         return WhiteTurn(board)
     }
